Delete server and rename simple_app -> app.sjs
